add remove unit tests
There are 2 unit tests for the non-existent element because the remove function in Mojave doesn’t check if an element exists and then fails shortly after trying to get that element’s parentNode.
The second unit test has that check manually added.

I’m not entirely sure how to check if this function did what it did. So I just used find() to see if we can find it. I feel like we should add another check before the removal function to see if it existed in the first place.
As of right now, that’s not really an issue; as previously mentioned remove() fails when the element doesn’t exist. Once it stops failing it would probably make sense to add an additional check. Or ask for a return in the function.

Signed-off-by: Patrick Kontschak <pk@becklyn.com>